clanker: Tighten ACP Connection.session_counter from u64 to u32 [imp-1787115836535037313]

+++ src/acp/server.zig
@@ -18,7 +18,7 @@ const Request = struct {
 
 pub const Connection = struct {
     initialized: bool = false,
-    session_counter: u64 = 0,
+    session_counter: u32 = 0,
     sessions: std.StringArrayHashMapUnmanaged(void) = .empty,
     prompt_busy: std.StringArrayHashMapUnmanaged(bool) = .empty,
